requirements.txt
python-jobspy
streamlit
crosslinked